move project delete button to project manage page
