+++ web/src/pages/projects/ManageProject.module.css
@@ -202,6 +202,49 @@
   gap: 1rem;
 }
 
+.delete-dialog {
+  padding: 1rem 0.5rem;
+  width: 100%;
+  max-width: 600px;
+  margin: 0 auto;
+}
+
+.delete-dialog h2 {
+  margin-top: 0;
+  margin-bottom: 1.5rem;
+  font-size: 1.5rem;
+  font-weight: 600;
+  color: #1a1a1a;
+  text-align: center;
+}
+
+.delete-dialog p {
+  margin-bottom: 1rem;
+  line-height: 1.6;
+  color: #1a1a1a;
+}
+
+.delete-dialog ul {
+  margin: 1rem 0 1.5rem 1.5rem;
+  padding: 0;
+  list-style-type: disc;
+}
+
+.delete-dialog li {
+  margin-bottom: 0.5rem;
+  color: #1a1a1a;
+  line-height: 1.5;
+}
+
+.delete-dialog strong {
+  color: #000;
+  font-weight: 600;
+}
+
+.delete-dialog button {
+  margin: 0.25rem;
+}
+
 @media (max-width: 768px) {
   .field {
     grid-template-columns: 1fr;
